+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Released on ?.
 - Fixed possible NRE in the markup formatters / attribute serialization (#903)
 - Fixed `GetSelector` returning invalid ID selector (#909)
 - Fixed `GetSelector` returning duplicate element ids (#910)
-- Fixed possible NRE in media elements without `src` (#914)
+- Fixed possible NRE in media and `embed` elements without `src` (#914)
 
 # 0.14.0